gpu version

+++ Models.py
@@ -38,7 +38,7 @@ class CNN(nn.Module):
         super().__init__(*args, **kwargs)
         self.is_coco = is_coco
         points_num = coco_point_num if self.is_coco else halpe_point_num
-        self.input_size = (points_num + box_feature_num / 2, 2)
+        self.input_size = int(points_num + box_feature_num / 2, 2)
         self.output_size = action_class_num if action_recognition else attitude_class_num
         self.Conv = nn.Sequential(
             nn.Conv2d(1, 3, kernel_size=3, padding=1),
